Disabled rate check in IRC class
The rate check is already in the message class to avoid authenticating
when the rate limit has already been reached, so this isn't necessary
for functionality. BUT it seems illogical to remove the rate limit check
in the class that actually does the rate limiting, so this will
hopefully go back in if I change the always-reconnect behavior.