Make hotkey ID names unique
Previously all hotkeys had the same internal name in C and were identified in the script by the pointer. This change appends the position to the string so the C names are unique. This allows tools such as OBSCommand to trigger the hotkeys remotely by name. The appended version of the name also matches the format of the identifiers used for saving the plugin data, so +1 for consistency.

Thank you @PaulBM for bringing this to my attention!